+++ xr_3da/EventAPI.h
@@ -7,7 +7,7 @@ typedef CEvent*		EVENT;
 class ENGINE_API CEventBase
 {
 public:
-	virtual void	OnEvent(EVENT E, DWORD P1, DWORD P2) = 0;
+	virtual void	OnEvent(EVENT E, u64 P1, u64 P2) = 0;
 };
 //---------------------------------------------------------------------
 class ENGINE_API CEventAPI
@@ -15,8 +15,8 @@ class ENGINE_API CEventAPI
 	struct Deferred
 	{
 		EVENT		E;
-		DWORD		P1;
-		DWORD		P2;
+		u64		P1;
+		u64		P2;
 	};
 private:
 	vector<EVENT>		Events;
@@ -29,10 +29,10 @@ public:
 	EVENT	Handler_Attach(const char* N, CEventBase* H);
 	void	Handler_Detach(EVENT& E, CEventBase* H);
 
-	void	Signal	(EVENT	E, DWORD P1=0, DWORD P2=0);
-	void	Signal	(LPCSTR E, DWORD P1=0, DWORD P2=0);
-	void	Defer	(EVENT	E, DWORD P1=0, DWORD P2=0);
-	void	Defer	(LPCSTR E, DWORD P1=0, DWORD P2=0);
+	void	Signal	(EVENT	E, u64 P1=0, u64 P2=0);
+	void	Signal	(LPCSTR E, u64 P1=0, u64 P2=0);
+	void	Defer	(EVENT	E, u64 P1=0, u64 P2=0);
+	void	Defer	(LPCSTR E, u64 P1=0, u64 P2=0);
 
 	void	OnFrame	();
 	void	Dump	();
