+++ xr_3da/xrGame/HUDManager.cpp
@@ -131,7 +131,7 @@ void CHUDManager::Render_Direct	()
 }
 
 //--------------------------------------------------------------------
-void CHUDManager::OnEvent(EVENT E, DWORD P1, DWORD P2)
+void CHUDManager::OnEvent(EVENT E, u64 P1, u64 P2)
 {
 }
 //--------------------------------------------------------------------
