+++ xr_3da/xrGame/Rain.h
@@ -78,7 +78,7 @@ private:
 public:
 	virtual void		OnDeviceCreate	();
 	virtual void		OnDeviceDestroy	();
-	virtual void		OnEvent			(EVENT E, DWORD P1, DWORD P2);
+	virtual void		OnEvent			(EVENT E, u64 P1, u64 P2);
 
 	void				Render			();
